requirements.txt
opencv-python
torch
numpy
ultralytics
torchvision 
scipy
opencv-python
torch
torchvision
numpy
scipy
pillow